Prototype Rails command
