+++ lib/retest/command.rb
@@ -66,7 +66,7 @@ module Retest
     end
 
     def rails_command
-      full_suite? ? ALL_RAILS_COMMAND : RAILS_COMMAND
+      Rails.command(all: full_suite?)
     end
 
     def rake_command
@@ -93,6 +93,25 @@ module Retest
         return 'bin/rspec' if bin_file
 
         'bundle exec rspec'
+      end
+    end
+
+    Rails = Struct.new(:all, :bin_file, keyword_init: true) do
+      def self.command(all:, bin_file: File.exist?('bin/rails'))
+        new(all: all, bin_file: bin_file).command
+      end
+
+      def command
+        return "#{root_command} <test>" unless all
+        root_command
+      end
+
+      private
+
+      def root_command
+        return 'bin/rails test' if bin_file
+
+        'bundle exec rails test'
       end
     end
   end
