+++ test/retest/command_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 require_relative 'command/auto_flag'
-require_relative 'command/rspec'
+require_relative 'command/setup_commands'
 
 module Retest
   class OptionsCommandTest < MiniTest::Test
